requirements.txt
flask
sentence-transformers
pdfminer.six
python-docx
numpy
